Teilnahme(zeit) kann gespeichert werden. Einschränkung der Stornierung.

+++ Classes/Controller/RegistrationController.php
@@ -10,9 +10,11 @@ use BoergenerWebdesign\BwRegistration\Utility\MailUtility;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Extbase\Error\Result;
 use TYPO3\CMS\Extbase\Mvc\Controller\ActionController;
+use TYPO3\CMS\Extbase\Mvc\Controller\MvcPropertyMappingConfiguration;
 use TYPO3\CMS\Extbase\Mvc\Exception\StopActionException;
 use TYPO3\CMS\Extbase\Mvc\Exception\UnsupportedRequestTypeException;
 use TYPO3\CMS\Extbase\Persistence\Exception\IllegalObjectTypeException;
+use TYPO3\CMS\Extbase\Persistence\Exception\UnknownObjectException;
 use TYPO3\CMS\Extbase\Persistence\Generic\PersistenceManager;
 use TYPO3\CMS\Extbase\Utility\DebuggerUtility;
 
@@ -163,13 +165,44 @@ class RegistrationController extends ActionController {
             $this -> addFlashMessage("Die Registrierung wurde erfolgreich storiert!");
             $this -> redirect('show', 'Event', null, ['event' => $registration -> getEvent(), 'slot' => $registration -> getSlot()]);
         } else if($registration && $registration -> getHash() == $hash) {
-            $this -> registrationRepository -> remove($registration);
-            $this -> mailUtility -> sendRevokeMail($registration);
+
+            if((new \DateTime()) >= $registration -> getSlot() -> getBeginDatetime()) {
+                $this -> view -> assign('error', 'Die Registrierung konnte nicht storniert werden, da die Veranstaltung bereits stattgefunden hat.');
+            } else if($registration -> isAttended()) {
+                $this -> view -> assign('error', 'Die Registrierung konnte nicht storniert werden, weil Sie bereits an der Veranstaltung teilgenommen haben.');
+            } else {
+                $this -> registrationRepository -> remove($registration);
+                $this -> mailUtility -> sendRevokeMail($registration);
+            }
 
             $this -> view -> assign('registration', $registration);
         }
     }
 
+    /**
+     * Passt das Eingabeformat der Zeitangaben an.
+     */
+    public function initializeUpdateAction() : void {
+        if($this -> request -> hasArgument('registration')) {
+            /** @var MvcPropertyMappingConfiguration $mappingConfiguration */
+            $mappingConfiguration = $this->arguments['registration']->getPropertyMappingConfiguration();
+            $mappingConfiguration -> setTargetTypeForSubProperty('attendedTime', 'array');
+        }
+    }
+
+    /**
+     * Aktualisiert eine Registrierung.
+     * @param Registration $registration
+     * @throws IllegalObjectTypeException
+     * @throws StopActionException
+     * @throws UnsupportedRequestTypeException
+     * @throws UnknownObjectException
+     */
+    public function updateAction(Registration $registration) : void {
+        $this -> registrationRepository -> update($registration);
+        $this -> addFlashMessage('Die Registrierung wurde aktualisiert.');
+        $this -> redirect('show', 'Event', null, ['event' => $registration -> getEvent(), 'slot' => $registration -> getSlot()]);
+    }
 
     /**
      * Tauscht die Indexe der Personen mit ihren temporären Uids aus.
